A MIPS Assembler
The Assembler will convert the basic MIPS insturctions(some of them) of any type(R/I/J Types) to its machine code.(in hex./bin.) A FILE of .asm will be taken as input and output with extension .obj should be generated. A project based on MIPS Assembly Language as its Assembler/Compiler.